+++ docs/snippets/installer/install-python.bash
@@ -23,9 +23,10 @@ function inst64_X_APP_NAME_X_install_with_pip() {
 
   bl64_msg_show_task 'deploy application'
   INST64_X_APP_NAME_CAPS_X_CLI_PATH="${HOME}/.local/bin/${INST64_X_APP_NAME_CAPS_X_CLI_NAME}"
+  # shellcheck disable=SC2086
   bl64_fs_set_umask "$BL64_FS_UMASK_RW_USER_RO_ALL" &&
     bl64_py_pip_usr_prepare &&
-    bl64_py_pip_usr_deploy "$INST64_X_APP_NAME_CAPS_X_PACKAGES"
+    bl64_py_pip_usr_deploy $INST64_X_APP_NAME_CAPS_X_PACKAGES
 }
 
 function inst64_X_APP_NAME_X_install_with_pipx() {
@@ -33,7 +34,8 @@ function inst64_X_APP_NAME_X_install_with_pipx() {
 
   bl64_msg_show_task 'deploy application'
   INST64_X_APP_NAME_CAPS_X_CLI_PATH="${INST64_X_APP_NAME_CAPS_X_CLI_NAME}"
-  "$INST64_X_APP_NAME_CAPS_X_PIPX_BIN" install "$INST64_X_APP_NAME_CAPS_X_PACKAGES"
+  # shellcheck disable=SC2086
+  "$INST64_X_APP_NAME_CAPS_X_PIPX_BIN" install $INST64_X_APP_NAME_CAPS_X_PACKAGES
 }
 
 # X_INSTALL_PLACEHOLDER_X