+++ test/bash/OEL/install-lazygit_01.sh
@@ -1,0 +1,1 @@
+../UB/install-lazygit_01.sh