+++ test/bash/OEL/install-pgadmin_01.sh
@@ -1,0 +1,1 @@
+../UB/install-pgadmin_01.sh