+++ test/bash/OEL/install-pipx_01.sh
@@ -1,0 +1,1 @@
+../UB/install-pipx_01.sh